do some renaming for visibility
